Updated version of Script 1
Script 1.3 has some improvements. The most important one is the correction of the CNN definition of number of parameters a bit better for the current case. This modification really improves the training!. Note that I have included (in comments) other possible modifications that, with the initial data set do not improve the training.